MdePkg/ShelLParametersLib: Add a shell parameter parsing library
TODO: It currently is backward compatible to old parsing capability
in ShellLib.
But it's not very clean.
Current lib when saying FlagTypeValue, it means the flag might be
followed by one value or none.
I am not sure supporting "none" is a valid requirement.
"reset -s [message] [-fwui]" requires this feature, so it's a valid
requirement.
"timezone -s -3:34" requires to treat "-3:34" as value of "-s".
So FlagTypeTimeValue is invented. But can we avoid inventing
this new flag type? Because if we need TimeValue flag type, do we
also need FlagTypeIntegerValue? This can avoid "NumberFirst" concept.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ruiyu Ni <ruiyu.ni@intel.com>