[Bug fix] Change PcdPciDisableBusEnumeration pcd type as dynamic
PciHostBridgeLib change this PCD value according to HOB, and which may be
used by other modules. So, this PCD type should be dynamic.

This is a buf fix for the below change in for_upstream branch:
UefiPayloadPkg: UefiPayload retrieve PCI root bridge from Guid Hob

When sending formal patch to ek2 commnunity, this two patch will be combined.

Signed-off-by: Zhiguang Liu <zhiguang.liu@intel.com>