UefiCpuPkg/PeiException: Fix HOOK_AFTER bug when two PEIMs link to it
When both PeimA and PeimB link to PeiCpuExceptionHandlerLib and
HOOK_AFTER is specified for exception #3.

The flow is as following:
1. PeimA installs Int3Handler to IDT[3]
2. PeimB installs Int3Handler to IDT[3] and saves the old IDT[3] to
   PeimB.OriginalInt3Handler
3. When INT#3 happens, PeimB.Int3Handler runs
    a. It saves stack content
    b. it updates stack.rip to PeimB.HookAfterInt3Handler
    c. It jumps to PeimB.OriginalInt3Handler
4. PeimA.Int3Handler (PeimB.OriginalInt3Handler) runs
    a. "IRETD" returns to PeimB.HookAfterInt3Handler
5. PeimB.HookAfterInt3Handler runs
    a. It restores the stack.rip
    b. "IRETD" returns to user's code

But since the global variable is saved in IDT[0].ExceptionHandlerData
When step #4 runs, it thinks it's running in step
address in user code.
PeimB.HookAfterInt3Handler doesn't run.

The patch adds a IDT[0].Previous field to point to the
previous IDT[0] so that PeimA in above case can find
its own global variable, instead of using the global data
of PeimB.

Change-Id: I1f48824dbbd83f0a31d36833a664191c63429222
Signed-off-by: Ray Ni <ray.ni@intel.com>