PcAtChipsetPkg/HpetTimerDxe: Fix nested interrupt time accuracy
When HPET timer is used as the timer interrupt and nested
interrupts into the HPET timer interrupt handler occur, the
elapsed time passed into the DXE Core is sometime too large
and this causes the DXE Core internal system time to run too
fast. Fix the logic so the previous main counter value stored
in the module global variable mPreviousMainCounter is always
captured before the timer notification function is called.

Without this change, mPreviousMainCounter is updated after
the timer notification function is called and when nesting
occurs, it updates with the value from the first level of
nesting which is further back in time than the interrupt from
the deepest level of nesting.  This causes the next timer
interrupt to compute a TimerPeriod that is larger than the
actual time period since the last interrupt and this cuases
the DXE Core sense of time to run too fast.

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>