UefiCpuPkg: simplify code logic in RestorePageTableAbove4G
In this commit, we remove input parameter IsValidPFAddress
to simplify the code logic in RestorePageTableAbove4G.

Previously, when SMM profile is enabled, the SMM page table
after ReadyToLock only covers the following 3 parts:
    1.SMRAM range
    2.SMM profile range
    3.MMIO range below 4G
So the access to MMIO range above 4G will also trigger PF
but should not be recorded. This results in that
SmmProfilePFHandler needs to distinguish whether the PF
address is within the MMIO range when PFAddress is above
4G. The IsValidPFAddress of RestorePageTableAbove4G
is needed to pass the information to SmmProfilePFHandler.

After we always creat full mapping SMM page table, the SMM
page table after ReadyToLock covers:
    1.SMRAM range
    2.SMM profile range
    3.MMIO range below 4G and above 4G
With the implementaion, when SMM profile is enabled, the
access to MMIO range above 4G won't go into the
SmmProfilePFHandler anymore(PF caused by code execution
in MMIO will be catched before SmmProfilePFHandler).
Then parameter IsValidPFAddress of RestorePageTableAbove4G
can be removed.

Signed-off-by: Dun Tan <dun.tan@intel.com>