StandaloneMmPkg/Core: Fix FV HOB loop by advancing with GET_NEXT_HOB
Previously, MmDispatchFvs could enter an infinite loop if a Firmware Volume HOB
with zero length was encountered, because GetNextHob() was called with the same
FvHob pointer repeatedly. This patch fixes the issue by passing
GET_NEXT_HOB(FvHob) to GetNextHob(), ensuring the loop advances to the next HOB
and preventing hangs.

Signed-off-by: Ray Ni <ray.ni@intel.com>